@@ -1,17 +1,16 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+import ReactDOM from 'react-dom/client'
 import App from './App.tsx'
 import './index.css'
 import { SocketContextProvider } from './context/SocketContext.tsx'
 import { GameContextProvider } from './context/GameContext.tsx'
 
-ReactDOM.render(
+ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <SocketContextProvider>
       <GameContextProvider>
         <App />
       </GameContextProvider>
     </SocketContextProvider>
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 )
